refactor(http): define http responses
